backend/requirements.txt
Flask
flask-cors
yfinance
anthropic
python-dotenv
